requirements.txt
langchain
streamlit
PyPDF2
python-dotenv